Move packages and remove unused code
