--- app/src/main/java/io/tvdubs/copixelate/data/TextField.kt
@@ -1,8 +1,0 @@
-package io.tvdubs.copixelate.data
-
-enum class TextField {
-    USER_EMAIL,
-    USER_PASSWORD,
-    USER_CONFIRM_PASSWORD,
-    USER_USERNAME
-}
