+++ app/src/main/java/io/tvdubs/copixelate/nav/ScreenInfo.kt
@@ -2,8 +2,5 @@ package io.tvdubs.copixelate.nav
 
 sealed class ScreenInfo(val route: String) {
     object Art: ScreenInfo(route = "art")
-    object Login: ScreenInfo(route = "login")
-    object Registration: ScreenInfo(route = "registration")
-    object Messages: ScreenInfo(route = "messages")
-    object MessageThread: ScreenInfo(route = "message_thread")
+    object Overview: ScreenInfo(route = "messages")
 }
